requirements.txt
scikit-learn==1.1.3
tensorflow==2.9.2
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-gcs-config==2.9.1
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-metadata==1.11.0
tensorflow-probability==0.17.0
streamlit==1.15.2
numpy==1.23.5
pandas==1.4.4
Pillow==9.2.0
gTTS==2.3.0
imageio==2.22.1
keras
Keras-Preprocessing==1.1.2
requests==2.28.1
google-cloud-storage==2.5.0
h5py==3.7.0
gcsfs
